bug fix: the program can only accept int type even if the method parameter type is double

+++ src/CoreCmd/Commands/MyTestCommand.cs
@@ -7,7 +7,7 @@ namespace CoreCmd.Commands
 {
     public class MyTestCommand
     {
-        public void MyTestMethod(string param1, int param2)
+        public void MyTestMethod(string param1, double param2)
         {
             Console.WriteLine("MyTestMethod: param1 = {0}; param2 = {1}", param1, param2);
         }
